modified:   examples/test.cc 	modified:   kernels/add.cl.c 	modified:   sources/Context.cc

+++ kernels/add.cl.c
@@ -3,6 +3,6 @@ void kernel add(
 		global const float* a, 
 		global const float* b)
 {
-	int i = get_global_id(0);
+	ulong i = get_global_id(0);
 	s[i] = a[i] + b[i];
 }
